+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Allow deleting graphs in the open from browser dialog
   * Show graph thumbnails and allow preview in the open from browser dialog
   * Create new empty graph
+  * Rename current graph
 
 ### Fixed
 * Ctrl-Y and Ctrl-Z descriptions are missing in the keyboard shortcuts help dialog. #90